requirements.txt
yt-dlp
requests
pyyaml
APScheduler
